requirements.txt
appdirs==1.4.3
astroid==2.4.2
bottle==0.12.19
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
importlib-metadata==3.4.0
ipaddr==2.2.0
isort==5.7.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pylint==2.6.0
pyparsing==2.4.6
pyserial==3.5
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
rope==0.18.0
six==1.14.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
zipp==3.4.0
